move agw specific backend plugin  code into its own file Signed-off-by: Joe McGuire <joe.alter.mcguire@gmail.com>
